Use the version catalog as the source of truth for plugin IDs
The IDs stated in the build-logic build script must exactly match the plugin
IDs in the version catalog file. This solution reduces the inherent
duplication.

To avoid any IDE errors about missing plugin accessors when adding new plugins
to the build-logic script, it's recommended to add the new ID in the version
catalog first, then invoke Gradle (sync) in order to generate the accessor
classes.